+++ typings/public.d.ts
@@ -1,5 +1,1 @@
-declare interface Lookup<T> {
-  [key: string]: T;
-}
-
 declare type int = number;